refactor: update retriever module for runtime compatibility
Move retriever to `src/lib/rag/`, fix import paths, simplify
AskMeChat, and implement fallback index for immediate runtime access.
